Spliting domoticz Module
