+++ Modules/pairingProcess.py
@@ -38,7 +38,7 @@ from Modules.orvibo import OrviboRegistration
 from Modules.configureReporting import processConfigureReporting
 from Modules.profalux import profalux_fake_deviceModel
 from Modules.logging import loggingHeartbeat, loggingPairing
-from Modules.domoticz import CreateDomoDevice
+from Modules.domoCreate import CreateDomoDevice
 from Modules.zigateConsts import HEARTBEAT, MAX_LOAD_ZIGATE, CLUSTERS_LIST, LEGRAND_REMOTES, LEGRAND_REMOTE_SHUTTER, LEGRAND_REMOTE_SWITCHS, ZIGATE_EP
 
 from Classes.IAS import IAS_Zone_Management
